refactor(manager) : manager 디렉토리 생성
- LottoManager 디렉토리를 위한 manager 디렉토리 생성
- LottoManager를 manager 디렉토리로 이동
