+++ src/main/java/lotto/manager/LottoManager.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.manager;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
